+++ src/module/rolls/index.ts
@@ -1,3 +1,3 @@
 export { IronswornRoll } from './ironsworn-roll'
 export { IronswornPrerollDialog } from './preroll-dialog'
-export { IronswornRollMessage } from './chat-message'
+export { IronswornRollMessage } from './ironsworn-roll-message'
